[fix/#47] 회원 탈퇴할 때 로그아웃도 진행, 로그아웃 시 구글 로그아웃도 함께 진행

+++ lib/provider/api/user/user_api.dart
@@ -5,6 +5,7 @@ import 'package:ac_project_app/models/result.dart';
 import 'package:ac_project_app/models/user/detail_user.dart';
 import 'package:ac_project_app/models/user/user.dart';
 import 'package:ac_project_app/provider/api/custom_client.dart';
+import 'package:ac_project_app/provider/logout.dart';
 
 class UserApi {
   final client = CustomClient();
@@ -65,7 +66,12 @@ class UserApi {
   }
   
   Future<bool> deleteUser() async {
-    final result = await client.deleteUri('/users');
-    return result.when(success: (_) => true, error: (_) => false);
+    final logoutResult = await logoutWithoutPush();
+    if (logoutResult) {
+      final result = await client.deleteUri('/users');
+      return result.when(success: (_) => true, error: (_) => false);
+    } else {
+      return false;
+    }
   }
 }
